+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 1.2.0
+    - Adicionada configuração de `ValorMonetarioWatcher`
+        - É possível deixar o símbolo de Real
+        - É possível manter os zeros quando o campo é apagado em lote
+    - Adicionado Builder para criação de watchers de valor
+
 ## 1.1.1:
     - Corrigido bug de ArrayIndexOutOfBounds no `BoletoBancarioTextWatcher` após apagar em lote
     - Adicionado teste de regressão para o caso acima na JVM que será executado no Travis